autostack-engine/pyproject.toml
[project]
name = "autostack-engine"
version = "0.1.0"
description = "This is the engine for infrastructure provisioning of the project"
authors = [
    {name = "Sakaria Nghivafe", email = "sakarianghivafe@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13,<4.0"

dependencies = [
    "confluent-kafka>=2.11.1,<3.0.0",
    "uvicorn>=0.35.0,<0.36.0",
    "httpx>=0.28.1,<0.29.0",
    "fastapi>=0.116.1,<0.117.0",
    "jinja2 (>=3.1.6,<4.0.0)",
    "jsonschema (>=4.25.1,<5.0.0)",
    "docker (>=7.1.0,<8.0.0)",
    "strawberry-graphql (>=0.282.0,<0.283.0)",
    "peewee (>=3.18.2,<4.0.0)",
    "beanie (>=2.0.0,<3.0.0)",
    "motor (>=3.7.1,<4.0.0)",
    "pymongo (>=4.15.1,<5.0.0)",
    "better-exceptions (>=0.3.3,<0.4.0)",
    "structlog (>=25.4.0,<26.0.0)",
    "gitpython (>=3.1.45,<4.0.0)",
    "openai (>=2.6.0,<3.0.0)",
    "ollama (>=0.6.0,<0.7.0)",
    "websockets (>=15.0.1,<16.0.0)",
    "google-genai (>=1.46.0,<2.0.0)",
    
]

[project.scripts]
consumers = "autostack_engine.scripts.run_consumers:main"
gateway = "autostack_engine.scripts.run_gateway:main"
service = "autostack_engine.scripts.run_service:main"
all = "autostack_engine.scripts.run_all:main"
setup-database = "autostack_engine.scripts.setup_database:main"
setup-kafka = "autostack_engine.scripts.setup_kafka:main"
health-check = "autostack_engine.scripts.health_check:main"
migrate-database = "autostack_engine.scripts.migrate_database:main"
create-migration = "autostack_engine.scripts.create_migration:main"
seed-database = "autostack_engine.scripts.seed_database:main"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "autostack-engine"
package-mode = true


[tool.poetry.dependencies]
python = ">=3.13,<4.0"

[tool.poetry.group.dev.dependencies]
pytest = ">=8.3.3,<9.0.0"
pytest-asyncio = ">=0.24.0,<0.25.0"
black = ">=24.8.0,<25.0.0"
isort = ">=5.13.2,<6.0.0"
flake8 = ">=7.1.1,<8.0.0"
mypy = ">=1.12.0,<2.0.0"
pytest-cov = ">=5.0.0,<6.0.0"

[tool.black]
line-length = 88
target-version = ['py313']

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.13"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
addopts = "--cov=services --cov=utils --cov-report=html"